[CON-2803] Unassign remote articles from local category

+++ Components/ImportService.php
@@ -121,32 +121,54 @@ class ImportService
      */
     public function unAssignArticleCategories(array $articleIds)
     {
-        // cast all items in $articleIds to int
-        // before use them in WHERE IN clause
-        foreach ($articleIds as $key => $articleId) {
-            $articleIds[$key] = (int)$articleId;
-        }
+        if (!empty($articleIds)) {
+            // cast all items in $articleIds to int
+            // before use them in WHERE IN clause
+            foreach ($articleIds as $key => $articleId) {
+                $articleIds[$key] = (int)$articleId;
+            }
 
+            $connection = $this->manager->getConnection();
+            $connection->beginTransaction();
+
+            try {
+                $attributeStatement = $connection->prepare(
+                    'UPDATE s_articles_attributes SET connect_mapped_category = NULL WHERE articleID IN (' . implode(", ", $articleIds) . ')'
+                );
+                $attributeStatement->execute();
+
+                $categoriesStatement = $this->manager->getConnection()->prepare('DELETE FROM s_articles_categories WHERE articleID IN (' . implode(", ", $articleIds) . ')');
+                $categoriesStatement->execute();
+
+                $categoryLogStatement = $this->manager->getConnection()->prepare('DELETE FROM s_articles_categories_ro WHERE articleID IN (' . implode(", ", $articleIds) . ')');
+                $categoryLogStatement->execute();
+                $connection->commit();
+
+            } catch (\Exception $e) {
+                $connection->rollBack();
+                throw new \Exception($e->getMessage());
+            }
+        }
+    }
+
+    /**
+     * Collect remote article ids by given category id
+     *
+     * @param int $localCategoryId
+     * @return array
+     */
+    public function findRemoteArticleIdsByCategoryId($localCategoryId)
+    {
         $connection = $this->manager->getConnection();
-        $connection->beginTransaction();
+        $sql = 'SELECT sac.articleID
+            FROM s_articles_categories sac
+            LEFT JOIN s_articles_attributes saa ON sac.articleID = saa.articleID
+            WHERE sac.categoryID = :categoryId AND saa.connect_mapped_category = 1';
+        $rows = $connection->fetchAll($sql, array(':categoryId' => $localCategoryId));
 
-        try {
-            $attributeStatement = $connection->prepare(
-                'UPDATE s_articles_attributes SET connect_mapped_category = NULL WHERE articleID IN (' . implode(", ", $articleIds) . ')'
-            );
-            $attributeStatement->execute();
-
-            $categoriesStatement = $this->manager->getConnection()->prepare('DELETE FROM s_articles_categories WHERE articleID IN (' . implode(", ", $articleIds) . ')');
-            $categoriesStatement->execute();
-
-            $categoryLogStatement = $this->manager->getConnection()->prepare('DELETE FROM s_articles_categories_ro WHERE articleID IN (' . implode(", ", $articleIds) . ')');
-            $categoryLogStatement->execute();
-            $connection->commit();
-
-        } catch (\Exception $e) {
-            $connection->rollBack();
-            throw new \Exception($e->getMessage());
-        }
+        return array_map(function($row) {
+            return $row['articleID'];
+        }, $rows);
     }
 
     /**
